Address @iroqueta review
